# Optimize cluster metadata creation