Metadata creation is massive contributor to pilot CPU utilization.
Currently, building the metadata takes ~15% of pilot CPU. An extra ~10%
is also hidden in MarshalToAny, as marshalling it is super slow as well.

A big chunk of this is in AddSubsetToMetadata. This does a proto.Clone
on the existing metadata. This is pretty wasteful, as we are already
creating a full copy of the metadata when we build the cluster.

Instead of a full copy, we can just insert the subset metadata (along
with the initial DR meta).

Results are 20% improvement in CPU.

```
name                     old time/op        new time/op        delta
ClusterGeneration/tls-6        26.7ms ± 2%        21.3ms ± 1%  -20.26%  (p=0.000 n=9+9)

name                     old kb/msg         new kb/msg         delta
ClusterGeneration/tls-6           813 ± 0%           813 ± 0%     ~     (all equal)

name                     old resources/msg  new resources/msg  delta
ClusterGeneration/tls-6         1.21k ± 0%         1.21k ± 0%     ~     (all equal)

name                     old alloc/op       new alloc/op       delta
ClusterGeneration/tls-6        9.63MB ± 0%        8.21MB ± 0%  -14.75%  (p=0.000 n=10+10)

name                     old allocs/op      new allocs/op      delta
ClusterGeneration/tls-6          152k ± 0%          126k ± 0%  -17.36%  (p=0.000 n=10+10)
```



[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[ ] Networking
[ ] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure


Pull Request Attributes

Please check any characteristics that apply to this pull request. 

[ ] Does not have any changes that may affect Istio users.